organize node optimization scheduling tests
